Refactor create login and run
