+++ src/index.ts
@@ -23,10 +23,6 @@
 import 'connect';
 
 import * as commander from 'commander';
-import {
-  login,
-  run,
-} from './commands';
 import apis from './commands/apis';
 import clone from './commands/clone';
 import create from './commands/create';
@@ -35,11 +31,13 @@ import deploy from './commands/deploy';
 import deployments from './commands/deployments';
 import help from './commands/help';
 import list from './commands/list';
+import login from './commands/login';
 import logout from './commands/logout';
 import logs from './commands/logs';
 import openCmd from './commands/openCmd';
 import pull from './commands/pull';
 import push from './commands/push';
+import run from './commands/run';
 import setting from './commands/setting';
 import status from './commands/status';
 import undeploy from './commands/undeploy';
